Corrected types of hardware properties.

+++ src/run/interface.py
@@ -62,13 +62,13 @@ device = 'cuda' if torch.cuda.is_available() else 'cpu'
 with st.expander('Hardware Properties'):
     def_mem = float(get_value(config, 'BENCHMARK', 'memory', 8))
     def_disk = float(get_value(config, 'BENCHMARK', 'disk', 500))
-    def_cores = float(get_value(config, 'BENCHMARK', 'cores', 8))
+    def_cores = int(get_value(config, 'BENCHMARK', 'cores', 8))
     memory = st.slider(
-        'Main Memory (GB)', min_value=1, 
-        max_value=1000, value=def_mem)
+        'Main Memory (GB)', min_value=1.0, 
+        max_value=1000.0, value=def_mem)
     disk = st.slider(
-        'Disk Space (GB)', min_value=1, 
-        max_value=10000, value=def_disk)
+        'Disk Space (GB)', min_value=1.0, 
+        max_value=10000.0, value=def_disk)
     cores = st.slider(
         'Number of Cores', min_value=1,
         max_value=1024, value=def_cores)
